Split test_agent_handoff_emitter.py into 7 concern-based files (#562)
Refactor pact-plugin/tests/test_agent_handoff_emitter.py (2231 lines,
17 classes, 108 collected tests) into 7 concern-based files plus a
conftest.py extraction of universal helpers. Zero behavior change;
parity preserved via pytest --collect-only node-ID diff (empty).

Files (max 598 lines each, all under 600 threshold):
- test_emitter_happy_and_gates.py    (322) — emit-at-all + 4 gates
- test_emitter_idempotency.py        (386) — marker-side dedup family
- test_emitter_path_sanitization.py  (598) — path-component security guards
- test_emitter_race_regression.py    (296) — #551 race-shape regression
- test_emitter_real_disk.py          (282) — empirical platform-shape suite
- test_emitter_resolution.py         (175) — name resolution + fallback
- test_emitter_robustness.py         (233) — malformed/null/exception input

conftest.py picks up VALID_HANDOFF + _run_main (universal across all 17
classes per consumption audit). PLATFORM_STDIN_SHAPE and _write_task_json
stay co-located in test_emitter_real_disk.py (single-file consumers).

Concern-based grouping chosen over fixture-shape: 16-of-17 mock-based
distribution makes a fixture-shape axis non-viable (one over-threshold
file plus one tiny extraction). Architecture rationale at
docs/architecture/562-emitter-test-split.md (worktree-only, gitignored).

Closes #562
